Fixed bezier drawing
If you take 5 line segments of each guide line you'll end up with 6
lines connecting each of those segments: 5 lines to connect the start of
each segment plus 1 for the end of the final segment.

I was calling repeat(...) with one argument and ended up getting three
elements. I'm still not sure how.

I upped the segments to 100 and now the animation makes sense.

Next step is to generate a bezier curve and use it as a guideline.
Optionally I may just have a bezier function that takes two lines and a
percentage and returns the point at that percentage along the
bezier.