Start of bezierBreakdowns test, list funs, tests
Started on the test for the bezier breakdowns that returns all the
computed lines and points for a list of guide lines and a single
fraction.

Added list functions equal, apply2 and all. Added some more tests and
converted a bunch of tests to use equal instead of laboriously checking
each array value and the array length.